move wizards to a pages/ folder separate from components (#14)
