+++ src/app/modules/onboarding/pages/hd-wallet-wizard/hd-wallet-wizard.component.ts
@@ -1,14 +1,16 @@
-import { Component, OnInit, ViewChild, NgZone, Input } from '@angular/core';
+import { Component, OnInit, ViewChild, Input } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, FormControl, AbstractControl } from '@angular/forms';
-import { MatStepper } from '@angular/material/stepper';
+import { Router } from '@angular/router';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
-import { tap, takeUntil, catchError, take, switchMap } from 'rxjs/operators';
+import { MatStepper } from '@angular/material/stepper';
+import { MatSnackBar } from '@angular/material/snack-bar';
+
+import { tap, takeUntil, catchError, switchMap } from 'rxjs/operators';
 import { Subject, throwError } from 'rxjs';
+
 import { WalletService, CreateWalletRequest, KeymanagerKind } from 'src/app/modules/core/services/wallet.service';
 import { MnemonicValidator } from '../../validators/mnemonic.validator';
-import { Router } from '@angular/router';
 import { AuthenticationService } from 'src/app/modules/core/services/auth.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-hd-wallet-wizard',
